communigate_cli-web/requirements.txt
PyYAML
Flask
ldap3
